requirements.txt
requests
beautifulsoup4
python-telegram-bot==20.5
PyMuPDF
tqdm
regex
